requirements.txt
setuptools
pyrogram
tgcrypto
asyncio
aiohttp
python-dotenv
tinydb
dnspython
pyfiglet
rich
requests
parsel
asyncio
progress
pytz